add static membership, read-committed isolation, regex subscription, deserializers
- ConsumerOptions.groupInstanceId sends the static identity through
  JoinGroup v3 / SyncGroup v3 / Heartbeat v3 / LeaveGroup v3 (KIP-345)
  with correct field order and throttle-time-aware response parsing;
  dynamic members keep legacy protocol versions
- isolationLevel: 'read_committed' sets Fetch isolation level and filters
  aborted transaction ranges until commit control markers (KIP-98)
- RecordSetDecoder understands transaction control records and producer ids
- encodeRecordBatch supports explicit base offsets and control attributes
- subscribe() accepts RegExp topic patterns expanded via cluster metadata
- keyDeserializer/valueDeserializer transform delivered messages

+++ test/chaos.test.ts
@@ -144,7 +144,7 @@ describe("deterministic Kafka chaos", () => {
       const held = (await consumer.fetch({ maxWaitMs: 1, maxMessages: 1 }))[0]!;
       await rejectsQuickly(consumer.fetch({ maxWaitMs: 1, maxMessages: 1 }));
       expect(partial).toBe(true);
-      expect(new TextDecoder().decode(held.value!)).toBe("value-0");
+      expect(new TextDecoder().decode(held.value! as Uint8Array)).toBe("value-0");
       expect(consumer.position(topic, 0)).toBe(1n);
     } finally { await client.disconnect(); close(broker); }
   });
